+++ src/main/kotlin/com/github/olivernybroe/wingidea/ide/highlighting/WingSyntaxHighlighter.kt
@@ -2,7 +2,6 @@ package com.github.olivernybroe.wingidea.ide.highlighting
 
 import com.github.olivernybroe.wingidea.lang.lexer.WingLexer
 import com.github.olivernybroe.wingidea.lang.psi.*
-import com.intellij.lexer.Lexer
 import com.intellij.openapi.editor.colors.TextAttributesKey
 import com.intellij.openapi.fileTypes.SyntaxHighlighterBase
 import com.intellij.psi.tree.IElementType
@@ -14,7 +13,7 @@ import com.github.olivernybroe.wingidea.lang.psi.WingElementTypes
  */
 class WingSyntaxHighlighter: SyntaxHighlighterBase() {
 
-    override fun getHighlightingLexer(): Lexer = WingLexer()
+    override fun getHighlightingLexer() = WingLexer()
 
     override fun getTokenHighlights(tokenType: IElementType?): Array<TextAttributesKey> =
         pack(attributes[tokenType])