+++ CHANGELOG.md
@@ -5,6 +5,12 @@
 ## [Unreleased]
 
 ### Added
+- Added folding suppport for blocks
+- Added folding suppport for imports
+
+## [0.0.2] - 2023-07-27
+
+### Added
 - Added support for caret position in arguments
 - Added brace matching support
 - Added commenter support
@@ -19,5 +25,6 @@
 ### Added
 - Initial scaffold created from [IntelliJ Platform Plugin Template](https://github.com/JetBrains/intellij-platform-plugin-template)
 
-[Unreleased]: https://github.com/olivernybroe/Wing-Idea/compare/v0.0.1...HEAD
+[Unreleased]: https://github.com/olivernybroe/Wing-Idea/compare/v0.0.2...HEAD
+[0.0.2]: https://github.com/olivernybroe/Wing-Idea/compare/v0.0.1...v0.0.2
 [0.0.1]: https://github.com/olivernybroe/Wing-Idea/commits/v0.0.1